backend/requirements.txt
Django==4.2.5
djangorestframework==3.14.0
django-filter==23.2
django-cors-headers==3.14.0
psycopg2-binary==2.9.7
